Slice 1: move shared match-query helpers out of the matches router
app/dashboard.py imported seven helpers straight out of the matches router,
against api/CLAUDE.md's "don't import another router's internals" rule. More
urgently, it blocked the next step: a match-details module needs four of those
same helpers, so importing them from the router while the router imports it back
is a circular import.

Move the shared query builders and loaded-row readers into a neutral
app/match_queries.py and repoint the dashboard. Nothing imports the router's
internals any more (only main.py, for the router itself).

Pure move: function bodies unchanged, no query shapes touched, and the OpenAPI
document is byte-identical, so schema.d.ts and Types.swift cannot have drifted.

Refs #195.
