+++ api/app/dashboard.py
@@ -10,16 +10,15 @@ from sqlalchemy.orm import selectinload
 
 from app.attention import attention_priority, list_attention_kind
 from app.db import get_session
-from app.matches import (
+from app.match_queries import (
     _attention_matches_query,
     current_game_number,
     match_eager_options,
     my_standing_proposal_exists,
     opponent_username,
     participant_filter,
-    side_win_counts,
 )
-from app.matches import (
+from app.match_queries import (
     my_side as resolve_my_side,
 )
 from app.models import (
@@ -33,6 +32,7 @@ from app.models import (
     UserLeagueRating,
 )
 from app.ratings import RatingStrategyKey, parse_strategy_key
+from app.result_acceptance import side_win_counts
 from app.retirement import retirement_deadline
 from app.schemas.dashboard import (
     AttentionKind,
